Added audio system

+++ include/drivers/rtl8139.h
@@ -4,6 +4,6 @@
 #include "../include/kernel/io/pci.h"
 
 void rtl8139_init(pci_dev_t dev);
-void rtl8139_send(uint16_t *data, uint32_t len);
+void rtl8139_send(uint8_t *data, uint32_t len);
 
 #endif
